fix a leak in procmesh.
Note that on current Cray with shared MPI library, we still
run out of MPI_Comm or something similar and crash in a call
to MPI_Comm_size in pfft. This happens after planning for about
120 ~ 256 FFTs. I haven't measure this precisely.

Also notice that on Edison.nersce.gov, the intel compiler produces
bad code for pfft and fftw. use the GNU compiler, which is also the
compiler they use to compiler fftw.