disable avx by default.
We need a more elegant way of dealing with these things.

A proper way is to add a command in setup.py that builds the pfft,
and declare mpicc, ldshared, and avx/sse as options. Then they can be read
in from a setup.cfg file.

There must also be a way to depend build_ext on build_pfft.